Including more fixes where the config param was missing
